Move all autocomplete from base to single place
