+++ base/views/learning_units/update.py
@@ -23,15 +23,11 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from dal import autocomplete
 from django.contrib.auth.decorators import login_required, permission_required
-from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db import IntegrityError
-from django.db.models import Q
 from django.http import HttpResponseRedirect, JsonResponse
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
-from django.utils.html import format_html
 from django.utils.translation import gettext_lazy as _
 from waffle.decorators import waffle_flag
 
@@ -39,12 +35,9 @@ from base.business import learning_unit_year_with_context
 from base.business.learning_units.edition import ConsistencyError
 from base.forms.learning_unit.edition import LearningUnitDailyManagementEndDateForm
 from base.forms.learning_unit.edition_volume import VolumeEditionFormsetContainer
-from base.forms.learning_unit.entity_form import find_additional_requirement_entities_choices
 from base.forms.learning_unit.learning_unit_postponement import LearningUnitPostponementForm
-from base.models.entity_version import find_pedagogical_entities_version
 from base.models.enums import learning_unit_year_subtypes
 from base.models.enums.entity_container_year_link_type import REQUIREMENT_ENTITIES
-from base.models.enums.organization_type import MAIN
 from base.models.learning_unit_year import LearningUnitYear
 from base.models.person import Person
 from base.models.proposal_learning_unit import ProposalLearningUnit
@@ -204,48 +197,3 @@ def _save_form_and_display_messages(request, form, learning_unit_year):
                           % {'year': e.last_instance_updated.academic_year})
         display_error_messages(request, e.error_list)
     return records
-
-
-class EntityAutocomplete(LoginRequiredMixin, autocomplete.Select2QuerySetView):
-    def get_queryset(self):
-        country = self.forwarded.get('country', None)
-        qs = find_additional_requirement_entities_choices()
-        if country:
-            qs = qs.exclude(entity__organization__type=MAIN).order_by('title')
-            if country != "all":
-                qs = qs.filter(entity__country_id=country)
-        else:
-            qs = find_pedagogical_entities_version().order_by('acronym')
-        if self.q:
-            qs = qs.filter(Q(title__icontains=self.q) | Q(acronym__icontains=self.q))
-        return qs
-
-    def get_result_label(self, result):
-        return format_html(result.verbose_title)
-
-
-class AllocationEntityAutocomplete(EntityAutocomplete):
-    def get_queryset(self):
-        self.forwarded['country'] = self.forwarded.get('country_allocation_entity')
-        return super(AllocationEntityAutocomplete, self).get_queryset()
-
-
-class AdditionnalEntity1Autocomplete(EntityAutocomplete):
-    def get_queryset(self):
-        self.forwarded['country'] = self.forwarded.get('country_additional_entity_1')
-        return super(AdditionnalEntity1Autocomplete, self).get_queryset()
-
-
-class AdditionnalEntity2Autocomplete(EntityAutocomplete):
-    def get_queryset(self):
-        self.forwarded['country'] = self.forwarded.get('country_additional_entity_2')
-        return super(AdditionnalEntity2Autocomplete, self).get_queryset()
-
-
-class EntityRequirementAutocomplete(EntityAutocomplete):
-    def get_queryset(self):
-        return super(EntityRequirementAutocomplete, self).get_queryset()\
-            .filter(entity__in=self.request.user.person.linked_entities)
-
-    def get_result_label(self, result):
-        return format_html(result.verbose_title)
